+++ app/src/main/java/hiof/gruppe15/treningsappen/ui/LoginScreen.kt
@@ -1,0 +1,40 @@
+package hiof.gruppe15.treningsappen.ui
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.Button
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun LoginScreen(login: () -> Unit) {
+    var password by remember { mutableStateOf("") }
+    Column(horizontalAlignment = Alignment.CenterHorizontally) {
+        OutlinedTextField(value = "Username", onValueChange = {})
+        OutlinedTextField(value = password,
+            onValueChange = { password = it },
+            visualTransformation = PasswordVisualTransformation(),
+            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password)
+        )
+
+        Button(onClick = { login() }) {
+            Text(text = "Logg inn")
+            
+        }
+
+    }
+}
+@Preview
+@Composable
+fun LoginPreview(){
+    LoginScreen(login = {})
+}
