+++ app/src/main/java/hiof/gruppe15/treningsappen/ui/WorkOutSession.kt
@@ -1,0 +1,31 @@
+package hiof.gruppe15.treningsappen.ui
+
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+
+
+@Composable
+fun WorkOutSession() {
+    Text(text = "you have just started your workout session",
+        textAlign = TextAlign.Center)
+}
+@Preview
+@Composable
+fun AppTopBarPreview(){
+    AppTopBar()
+}
+
+@Preview
+@Composable
+fun WorkOutSessionPreview(){
+    WorkOutSession()
+}
+@Preview
+@Composable
+fun BottomBarPreview() {
+    AppBottomBar()
+
+}
